ai-packages/rag-module/requirements.txt
apscheduler==3.11.0
fastapi==0.115.5
grpcio==1.68.1
grpcio-tools==1.68.1
langchain==0.3.15
langchain-cli==0.0.35
langchain-community==0.3.15
langchain-core==0.3.31
langchain-google-vertexai==2.0.11
langchain_ibm==0.3.5
langchain-ollama==0.2.2
langchain-openai==0.3.2
opensearch-py==2.8.0
python-dotenv==1.0.1
python-keycloak==4.7.3
requests==2.32.3
